# Mutual TLS for connection between proxy and bin