This generates a certificate that is then used to secure and mutually authenticate communication between the binary and the test. 

Interestingly, creating the proxy and the certificate generation is a very small part of the overhead when compared to https://github.com/lox/bintest/commit/67b6a20adb653219474b963c4519202ee12beecf:

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkCreatingProxies-4     477515503     432421627     -9.44%
BenchmarkCallingProxies-4      6294863       21302862      +238.42%

benchmark                     old MB/s     new MB/s     speedup
BenchmarkCallingProxies-4     9.53         2.82         0.30x

benchmark                      old allocs     new allocs     delta
BenchmarkCreatingProxies-4     247            1915           +675.30%
BenchmarkCallingProxies-4      337            2004           +494.66%

benchmark                      old bytes     new bytes     delta
BenchmarkCreatingProxies-4     82224         164592        +100.18%
BenchmarkCallingProxies-4      131131        263677        +101.08%
```